% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{biomass}
\alias{biomass}
\title{Biomass data}
\format{
A tibble with 192 rows and 7 variables:
\describe{
\item{year}{year of sampling}
\item{siteID}{unique site code}
\item{blockID}{unique block code}
\item{plotID}{unique plot code}
\item{treatment}{functional group removal treatment}
\item{removed_fg}{functional group removed}
\item{biomass}{biomass value in g}
...
}
}
\source{
\url{https://osf.io/4c5v2/}
}
\usage{
biomass
}
\description{
A dataset containing one year of biomass data from functional group removal experiment from four alpine sites in western Norway.
}
\keyword{datasets}
