% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_durin_metadata.R
\name{create_durin_meta_data}
\alias{create_durin_meta_data}
\title{Create durin meta data}
\usage{
create_durin_meta_data(study = NULL, csv_output = FALSE, filename = NULL)
}
\arguments{
\item{study}{select study for which metadata will be created. 4Corners: study on the 4 main Durin sites; DroughtNet: droughtNet experiment at Lygra and Tjotta; Nutrient: nutrient experiment at Lygra}

\item{csv_output}{logical; argument csv_output or not.
The default is csv_output = FALSE. If csv_output = TRUE a csv file is produced and saved in the project directory.}

\item{filename}{file name for storing csv output. Default is NULL. The path is only needed i csv_output = TRUE. The file name contains Durin_study_filename.csv, where the file name is defined by the user.}
}
\value{
a tibble and optionally a csv file
}
\description{
Function to make durin meta data.
}
\examples{
create_durin_meta_data()

}
