% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funder_meta_data.R
\name{create_funder_meta_data}
\alias{create_funder_meta_data}
\title{Create funder meta data}
\usage{
create_funder_meta_data(path = NULL, csv_output = FALSE)
}
\arguments{
\item{path}{path for storing csv output. Default is NULL. The path is only needed i csv_output = TRUE.}

\item{csv_output}{logical; argument csv_output or not.
The default is csv_output = FALSE. If csv_output = TRUE a csv file is produced and saved under the path name.}
}
\value{
a tibble and optionally a csv file
}
\description{
Function to make funder meta data.
}
\examples{
create_funder_meta_data()

}
