% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_threed_ZA_data_dic.R
\name{create_threed_ZA_meta_data}
\alias{create_threed_ZA_meta_data}
\title{Create ThreeD ZA meta data}
\usage{
create_threed_ZA_meta_data(csv_output = FALSE, filename = NULL)
}
\arguments{
\item{csv_output}{logical; argument csv_output or not.
The default is csv_output = FALSE. If csv_output = TRUE a csv file is produced and saved in the project directory.}

\item{filename}{file name for storing csv output. Default is NULL. The path is only needed i csv_output = TRUE. The file name contains ThreeD_filename.csv, where the file name is defined by the user.csv_output = TRUE a csv file is produced and saved under the path name.}
}
\value{
a tibble and optionally a csv file
}
\description{
Function to make ThreeD meta data for South Africa.
}
\examples{
create_threed_meta_data()

}
