% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{description_table}
\alias{description_table}
\title{Description table}
\format{
A tibble with 192 rows and 7 variables:
\describe{
\item{TableID}{unique code for each dataset}
\item{Variable name}{variable name}
\item{Description}{variable description}
\item{Units}{units for variables}
\item{How measured}{how a variable was measured}
...
}
}
\source{
\url{https://osf.io/4c5v2/}
}
\usage{
description_table
}
\description{
A table describing the dataset.
}
\keyword{datasets}
