% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dic}
\alias{dic}
\title{Block dictionary}
\format{
A tibble with 48 rows and 2 variables:
\describe{
\item{funder_blockID}{unique code for funder blockID}
\item{funcab_blockID}{unique code for funcab blockID}
...
}
}
\source{
\url{https://osf.io/tx9r2/}
}
\usage{
dic
}
\description{
A table block_dictionary.
}
\keyword{datasets}
