% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{funcab}
\alias{funcab}
\title{FunCaB}
\format{
A tibble with 4 rows and 3 variables:
\describe{
\item{funder_blockID}{unique code for funder blockID}
\item{var}{random variable}
\item{treatment}{unique code for treatement}
...
}
}
\usage{
funcab
}
\description{
A random FunCaB dataset.
}
\keyword{datasets}
