% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{previous_description_table}
\alias{previous_description_table}
\title{Previous description table}
\format{
A tibble with 2 rows and 5 variables:
\describe{
\item{TableID}{unique code for each dataset}
\item{Variable name}{variable name}
\item{Description}{variable description}
\item{Units}{units for variables}
\item{How measured}{how a variable was measured}
...
}
}
\usage{
previous_description_table
}
\description{
A minimal description table containing only year and siteID variables.
Used to demonstrate pre-filling in make_description_table().
}
\keyword{datasets}
