% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcabization.R
\name{funcabization}
\alias{funcabization}
\title{Convert funder terminology to funcab terminology and vice versa}
\usage{
funcabization(dat, convert_to = "FunCaB")
}
\arguments{
\item{dat}{data input file to convert}

\item{convert_to}{Conversion to desired terminology. FunCaB means that Funder terminology is covnerted to FunCaB, and else it is converted to Funder.}
}
\value{
a tibble
}
\description{
Function to Convert funder terminology to funcab terminology and vice versa.
}
\examples{
data(funder)
funcabization(dat = funder, convert_to = "FunCaB")
data(funcab)
funcabization(dat = funcab, convert_to = "Funder")

}
