% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_started_data_dic.R
\name{get_started}
\alias{get_started}
\title{Make data dictionary directory and description file}
\usage{
get_started()
}
\value{
a directory and a empty description table
}
\description{
Function to make a new directory for the data dictionary directory and a template for the description file. This function should be run first, before making a data dictionary. It only has to be run once.
}
\examples{
get_started()

}
